pyproject.toml
[project]
name = "books-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "llama-index-llms-openai>=0.2.0",
    "python-dotenv>=1.0.1",
    "llama-index>=0.11.3",
    "llama-index-llms-ollama>=0.3.0",
    "arize-phoenix>=4.33.0",
    "llama-index-callbacks-arize-phoenix>=0.2.1",
    "ebooklib>=0.18",
    "html2text>=2024.2.26",
    "ipykernel>=6.29.5",
    "flask>=3.0.3",
    "uvicorn>=0.30.6",
    "snakeviz>=2.2.0",
]
